agentic_search/pyproject.toml
[project]
name = "agentic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "google-search-results>=2.4.2",
    "lxml>=6.0.4",
    "pydantic>=2.12.5",
    "requests>=2.33.1",
    "tavily>=1.1.0",
    "uvicorn>=0.44.0",
]
